requirements.txt
Flask==2.3.3
gunicorn==21.2.0
cryptography==41.0.7
supabase==2.0.2
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
httpx>=0.24.0,<0.25.0
authlib==1.2.1
python-multipart==0.0.6
Werkzeug==2.3.7